GCS: Wait for eventual consistency before returning from various operations.
This should help spurious test failures (and probably some prod scenarios). This obviously isn't a complete solution - just because GCS says something is
true once doesn't mean it's going to stay true, but at least we have a higher chance of succeeding various ops.

I'd love for the actual solution for the tests to include a faked client (a la moto), but it doesn't seem like that's anywhere on the roadmap :(.