Allow batch delegates for batchable classes
This allows a class to specify another class that is responsible for
handling batched versions of tasks. This is useful when the method for
running in batch mode is significantly different from the method for
running a single job or when it would be too difficult to change the
parameter type of a job for batching.