Set datatable order correctly on hash change
We were setting order by overwriting the datatable's order function in
the processHashChange function in visualiserApp.js. This resulted in a
bug that would prevent tab changes when we try to look up the order by
calling the order function after it is replaced with an array.

The fix is pretty simple, as we just need to use the single-argument
order function that acts as a setter.