Update workers to check actual_complete on deps [D27686]
Workers are supposed to check that a job is ready to run before running
them because the scheduler state is not always 100% up-to-date.
Unfortunately, workers have not been checking the dirty jobs table,
resulting in inconsistent reruns if an undo is timed poorly. This fixes
the issue by calling actual_complete rather than complete in worker.py