Fix flake errors
Summary:
It looks like some new rules added in October 2018 have caused flake8 to
start failing in this repo. It's entirely about bad escaping in string
literals, so it's an easy fix.

Note that having this fail as new linting rules are added was
intentional, as we want the code to be high quality and readable. We
might want to revisit that decision now.

Test Plan: `tox -e flake8`

Reviewers: zviki, ben, jimmy.chen

Reviewed By: ben, jimmy.chen

Subscribers: #dw_review

Differential Revision: https://cr.houzz.net/D43478